requirements.txt
streamlit==1.26.0
requests==2.31.0
selenium==4.12.0
webdriver-manager==3.8.6
Pillow
beautifulsoup4==4.12.2
python-decouple
openai==0.28.0
python-dotenv
PyPDF2
speechrecognition
pydub
numpy
PyAudio
sentence-transformers
faiss-cpu
scikit-learn